move textColor to separate file
